python/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
opencv-python>=4.8.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
numpy>=1.24.0
Pillow>=10.0.0
pynvml>=11.0.0
demucs>=4.0.0
rembg[gpu]
